requirements.txt
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.2
numpy==1.19.0
pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.0
six==1.15.0
torchsummary==1.5.1
